[HGS] Move character select screen into a separate component
