+++ package.json
@@ -35,6 +35,7 @@
 	},
 	"type": "module",
 	"dependencies": {
+		"dexie": "^4.0.10",
 		"fuse.js": "^7.0.0",
 		"sqlite3": "^5.1.7"
 	}
